Acceptable for grade 3 without tests

+++ ticket-service/src/main/java/com/epam/training/ticketservice/commandprompt/commands/MovieCommand.java
@@ -2,12 +2,13 @@ package com.epam.training.ticketservice.commandprompt.commands;
 
 import com.epam.training.ticketservice.commandprompt.commands.checks.CheckService;
 import com.epam.training.ticketservice.core.movie.MovieService;
-import com.epam.training.ticketservice.core.movie.model.Movie;
+import com.epam.training.ticketservice.core.movie.model.MovieDto;
 import lombok.AllArgsConstructor;
 import org.springframework.shell.standard.ShellComponent;
 import org.springframework.shell.standard.ShellMethod;
 
-import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
 @ShellComponent
 @AllArgsConstructor
@@ -16,33 +17,38 @@ public class MovieCommand {
     private final CheckService checkService;
 
     @ShellMethod(key = "create movie",value = "Create a new movie")
-    public Movie createMovie(String title, String genre, Integer lengthInMinutes) throws Exception {
-        if (movieService.getMovieByTitle(title) != null) {
-            throw new Exception("There already is a movie with that title");
+    public String createMovie(String title, String genre, Integer lengthInMinutes) {
+        if (!movieService.getMovieList().isEmpty() && checkService.isExistingMovie(title)) {
+            return "There already is a movie with that title";
         }
-        Movie movie = new Movie(title,genre,lengthInMinutes);
-        movieService.createMovie(movie);
-        return movie;
+        MovieDto movieDto = new MovieDto(title,genre,lengthInMinutes);
+        movieService.createMovie(movieDto);
+        return movieDto.toString();
     }
 
     @ShellMethod(key = "update movie",value = "Update a movie")
-    public Movie updateMovie(String title,String genre, Integer lengthInMinutes) throws Exception {
-        checkService.checkMovie(title);
+    public String updateMovie(String title, String genre, Integer lengthInMinutes) {
+        if (!checkService.isExistingMovie(title)) {
+            return "There is no movie with that title";
+        }
         movieService.updateMovie(title, genre, lengthInMinutes);
-        return new Movie(title, genre, lengthInMinutes);
+        return new MovieDto(title, genre, lengthInMinutes).toString();
     }
 
     @ShellMethod(key = "delete movie",value = "Delete a movie")
-    public void deleteMovie(String title) throws Exception {
-        checkService.checkMovie(title);
+    public String deleteMovie(String title) {
+        if (!checkService.isExistingMovie(title)) {
+            return "There is no movie with that title";
+        }
         movieService.deleteMovie(title);
+        return title + " successfully deleted";
     }
 
     @ShellMethod(key = "list movies",value = "List of movies")
-    public void listMovies() throws Exception {
-        if (movieService.getMovieList() == null) {
-            throw new Exception("There are no movies at the moment");
+    public String listMovies() {
+        if (movieService.getMovieList().isEmpty()) {
+            return "There are no movies at the moment";
         }
-        movieService.getMovieList().forEach(System.out::println);
+        return movieService.getMovieList().stream().map(Objects::toString).collect(Collectors.joining("\n"));
     }
 }
